BoardServiceTest 작성 1차 완료

+++ src/main/java/com/hanbaguni/project/domain/user/service/BoardServiceImpl.java
@@ -9,15 +9,16 @@ import com.hanbaguni.project.domain.user.dto.BoardDto;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import static com.hanbaguni.project.global.utils.DtoUtils.isValidDto;
 import static com.hanbaguni.project.global.utils.EntityUtils.updateIfNotNull;
 
 @Service
@@ -33,6 +34,11 @@ public class BoardServiceImpl implements BoardService {
     public Board createBoard(String username, BoardDto boardDto) {
         Member newMember = memberRepository.findByUsername(username)
                 .orElseThrow(() -> new EntityNotFoundException("user not found."));
+
+        Optional<Board> findBoard = boardRepository.findByTitle(boardDto.getTitle());
+        if(findBoard.isPresent()){
+            throw new IllegalStateException("중복된 제목입니다");
+        }
 
         Board newBoard = Board.builder()
                 .member(newMember)
@@ -51,13 +57,18 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
-    public List<BasicBoardDto> getAllBoards(String username) {
+    public List<Board> getAllBoards() {
+        return boardRepository.findAll();
+    }
+
+    public List<BasicBoardDto> getAllMemberBoards(String username) {
         List<Board> boardList = memberRepository.findByUsername(username)
                 .orElseThrow(()->new EntityNotFoundException("user not found"))
                 .getBoards();
 
         return boardList.stream()
                 .map(board-> BasicBoardDto.builder()
+                        .id(board.getId())
                         .title(board.getTitle())
                         .link(board.getLink())
                         .staff(board.getStaff())
@@ -71,7 +82,19 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
-    public void deleteBoard(Long boardId) {
+    public Board getBoard(Long boardId) {
+        return boardRepository.findById(boardId)
+                .orElseThrow(()-> new EntityNotFoundException("board not found"));
+    }
+
+
+    @Override
+    public void deleteBoard(Long boardId,String username) {
+        Board board = boardRepository.findById(boardId)
+                .orElseThrow(()->new EntityNotFoundException("게시글을 찾을 수 없습니다."));
+        if(!Objects.equals(board.getMember().getUsername(), username)){
+            throw new AccessDeniedException("권한이 없습니다.");
+        }
         boardRepository.deleteById(boardId);
     }
